Refactor into gametest mod, convert all the framework into one line
Only loss is having to restart to add a new class, but it wasn't that
useful anyway (couldn't reload). And many things got better
