+++ src/gametest/java/carpetextra/ExportPathChanger.java
@@ -1,0 +1,11 @@
+package carpetextra;
+
+import net.fabricmc.api.ModInitializer;
+import net.minecraft.test.StructureTestUtil;
+
+public class ExportPathChanger implements ModInitializer {
+    @Override
+    public void onInitialize() {
+        StructureTestUtil.testStructuresDirectoryName = "../src/gametest/resources/data/carpet-extra/gametest/structure";
+    }
+}
