--- src/gametest/java/carpetextra/TestUtils.java
@@ -1,10 +1,0 @@
-package carpetextra;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-public class TestUtils {
-
-    public static final Logger LOGGER = LoggerFactory.getLogger("CarpetExtraTest");
-
-}
